Moved state_compressor setup to util/state_compressor.py
